change api: pyuart.uart()
