--- aelib/.gitignore
@@ -1,4 +1,0 @@
-__pycache__
-*~
-.*~
-
