+++ pypi/setup.py
@@ -5,14 +5,14 @@ from os import path
 here = path.abspath(path.dirname(__file__))
 
 
-with open(path.join(here, 'README.md'), encoding='utf-8') as f:
+with open(path.join(here, '../README.md'), encoding='utf-8') as f:
     long_description = f.read()
 
 setup(
     name='pyuart', # パッケージ名(プロジェクト名)
     packages=['pyuart'], # パッケージ内(プロジェクト内)のパッケージ名をリスト形式で指定
 
-    version='0.0.5', # バージョン
+    version='0.1.1', # バージョン
 
     license='MIT', # ライセンス
 
@@ -25,7 +25,7 @@ setup(
     url='https://github.com/arkedge/pyuart', # パッケージに関連するサイトのURL(GitHubなど)
 
     description='a simple UART library for python3', # パッケージの簡単な説明
-    long_description='pyuart', # PyPIに'Project description'として表示されるパッケージの説明文
+    long_description=long_description, # PyPIに'Project description'として表示されるパッケージの説明文
     long_description_content_type='text/markdown', # long_descriptionの形式を'text/plain', 'text/x-rst', 'text/markdown'のいずれかから指定
     keywords='pyuart uart', # PyPIでの検索用キーワードをスペース区切りで指定
