+++ sample.tx.py
@@ -1,9 +1,9 @@
 #!/usr/bin/env python3
 
-from aelib import uart
+from pyuart import uart
 
 if __name__ == '__main__':
-    u0 = uart.uart(dev='/dev/ttyUSB0', baudrate = 115200, timeout = 15)
+    u0 = uart(dev='/dev/ttyUSB0', baudrate = 115200, timeout = 15)
     u0.open()
     u0.diag()
